feat: Problem solved with optimization!
Add functions: max, allocate2DSizeTArray and longest_common_substring optimization

+++ src/main.c
@@ -4,7 +4,7 @@ int main()
 {
     const char *word_to_check = "Hish";
     const char *word_to_compare = "Fish";
-    const int result = longest_common_substring(word_to_check, word_to_compare);
-    printf("\nThe largest substring length is: %i\n", result);
+    const size_t result = longest_common_substring(word_to_check, word_to_compare);
+    printf("\nThe largest substring length is: %zu\n", result);
     return 0;
 }
